feat(Dynamics): type classes for FunLike monoids with nice composition
The type synonyms `Function.End`, `Monoid.End`, `Module.End`, etc. all
instance `Monoid` in the same manner, where the `mul` field is
composition (of bundled endomorphisms), `one` is the identity
endomorphism, and `mul_assoc` is the proof that function composition is
associative. Making a `Monoid` instance for each End type synonym in
this manner is not difficult. However, it is tedious to translate
algebraic properties of the monoid to set-theoretic properties of the
unbundled endomorphisms. The type classes `CompSemigroup` and its
extensions make this translation easier.